backstage: introduce script to determine what percy jobs to run
